+++ frontend/lib/lib.zakariClient.js
@@ -140,7 +140,8 @@ class ZakariClient {
 
   /**
    * @typedef {Object} SpellcheckMessage
-   * @property {number} status - The X Coordinate
+   * @property {number} id - The id of the stored result
+   * @property {number} status - The status of the result
    * @property {string} kreyol - The kreyol used
    * @property {Array<string>} unknown_words - List of unrecognized words in the request
    * @property {string} message - The spellchecked result
@@ -181,6 +182,47 @@ class ZakariClient {
             withCredentials: true,
           }
         )
+        .then(
+          (result) => {
+            const rep = result.data
+            resolve(rep)
+          },
+          (reason) => {
+            const data = me.handleFailure(reason)
+            return reject(data)
+          }
+        )
+        .catch((error) => {
+          return reject({ code: 500, status: 'error', error })
+        })
+    })
+  }
+
+  /**
+   * @typedef {Object} CorrectionRating
+   * @property {number} rating - a rating from 0 to 5
+   * @property {string} user_correction - a correction proposed by the user
+   * @property {string} user_notes - notes of comment on the rating fron the user
+   */
+
+  /**
+   * Rate the provided correction
+   * @param {string} msgId The id of the correction to rate
+   * @param {CorrectionRating} rating the rating info
+   */
+  rateCorrection(msgId, rating) {
+    const me = this
+    const url = `${me.host}/api/account/spellcheck/${msgId}/rating`
+    return new Promise((resolve, reject) => {
+      return axios
+        .post(url, rating, {
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: me._authorization,
+            Cookie: me._cookies.join('; '),
+          },
+          withCredentials: true,
+        })
         .then(
           (result) => {
             const rep = result.data