+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- correction can be rated
+
 ## [1.0.5] - 2022-05-17
 
 - call mailgun api directly